MSI-1707: Fixed source and stock source link dto's return types on getters

+++ app/code/Magento/InventorySales/Model/ResourceModel/GetAssignedStockIdForWebsite.php
@@ -48,6 +48,6 @@ class GetAssignedStockIdForWebsite implements GetAssignedStockIdForWebsiteInterf
         if (count($result) === 0) {
             return null;
         }
-        return reset($result);
+        return (int)reset($result);
     }
 }
